requirements.txt
httpx>=0.28,<0.29
python-jose[cryptography]>=3.5,<4
